# contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "InfraPulse MSK",
    "description": "API InfraPulse MSK. Документация открыта для чтения; каждый маршрут, кроме `/health/*` и входа, проверяет сессию или токен интеграции.\n\n* **Чтение (GET)** — по cookie сессии: войти в веб-интерфейс этого же сайта, затем «Try it out» здесь.\n* **Пачки наблюдений** `POST /api/v1/observations` — токен интеграции: «Authorize» → `integrationToken`, значение `ipk_...` без слова Bearer.\n* **Запись под сессией (POST)** — дополнительно `csrfToken`: значение cookie `__Host-infrapulse-csrf`.\n\nИнструкция для систем-источников — `docs/INTEGRATION_API.md` в репозитории.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/reports/monthly": {
      "get": {
        "summary": "Monthly",
        "operationId": "monthly_api_v1_reports_monthly_get",
        "parameters": [
          {
            "name": "month",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^\\d{4}-(0[1-9]|1[0-2])$",
              "title": "Month"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MonthlyReport"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "sessionCookie": []
          }
        ]
      }
    },
    "/api/v1/reports/monthly.xlsx": {
      "get": {
        "summary": "Monthly Xlsx",
        "operationId": "monthly_xlsx_api_v1_reports_monthly_xlsx_get",
        "parameters": [
          {
            "name": "month",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^\\d{4}-(0[1-9]|1[0-2])$",
              "title": "Month"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/vnd.openxmlformats-officedocument.spreadsheetml.sheet": {}
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "sessionCookie": []
          }
        ]
      }
    },
    "/api/v1/reports/journal.xlsx": {
      "get": {
        "summary": "Journal Xlsx",
        "operationId": "journal_xlsx_api_v1_reports_journal_xlsx_get",
        "parameters": [
          {
            "name": "issued_from",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date",
              "title": "Issued From"
            }
          },
          {
            "name": "issued_to",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date",
              "title": "Issued To"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/vnd.openxmlformats-officedocument.spreadsheetml.sheet": {}
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "sessionCookie": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "MonthlyReport": {
        "properties": {
          "mode": {
            "type": "string",
            "enum": [
              "fixture",
              "replay",
              "received",
              "live"
            ],
            "title": "Mode"
          },
          "month": {
            "type": "string",
            "pattern": "^\\d{4}-(0[1-9]|1[0-2])$",
            "title": "Month"
          },
          "period_start": {
            "type": "string",
            "format": "date",
            "title": "Period Start"
          },
          "period_end": {
            "type": "string",
            "format": "date",
            "title": "Period End"
          },
          "generated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Generated At"
          },
          "data_as_of": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Data As Of"
          },
          "basis": {
            "type": "string",
            "const": "automatic_registered_event",
            "title": "Basis",
            "default": "automatic_registered_event"
          },
          "cards": {
            "items": {
              "$ref": "#/components/schemas/ReportCardCounts"
            },
            "type": "array",
            "title": "Cards"
          },
          "decisions": {
            "items": {
              "$ref": "#/components/schemas/ReportDecisionCount"
            },
            "type": "array",
            "title": "Decisions"
          },
          "top_objects": {
            "items": {
              "$ref": "#/components/schemas/ReportObjectRow"
            },
            "type": "array",
            "maxItems": 10,
            "title": "Top Objects"
          },
          "alarms_by_type": {
            "items": {
              "$ref": "#/components/schemas/ReportAlarmTypeRow"
            },
            "type": "array",
            "title": "Alarms By Type"
          },
          "notes": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Notes"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "mode",
          "month",
          "period_start",
          "period_end",
          "generated_at"
        ],
        "title": "MonthlyReport"
      },
...
      "ReportAlarmTypeRow": {
        "properties": {
          "sensor_type": {
            "type": "string",
            "minLength": 1,
            "title": "Sensor Type"
          },
          "source_alarms": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Source Alarms"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "sensor_type",
          "source_alarms"
        ],
        "title": "ReportAlarmTypeRow"
      },
      "ReportCardCounts": {
        "properties": {
          "target_spec_id": {
            "type": "string",
            "enum": [
              "sensor-failure/connection_loss_1h",
              "sensor-failure/technical_value",
              "phase-feeders/phase_loss_all",
              "fire-detectors/failure_pilot"
            ],
            "title": "Target Spec Id"
          },
          "issued": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Issued"
          },
          "released": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Released"
          },
          "no_event": {
            "type": "integer",
            "minimum": 0.0,
            "title": "No Event"
          },
          "unknown": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Unknown"
          },
          "open": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Open"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "target_spec_id",
          "issued",
          "released",
          "no_event",
          "unknown",
          "open"
        ],
        "title": "ReportCardCounts",
        "description": "выдано = снята по событию + без события + неизвестно + открыто."
      },
      "ReportDecisionCount": {
        "properties": {
          "decision_code": {
            "type": "string",
            "enum": [
              "R1",
              "R2",
              "R3",
              "R4",
              "R5",
              "R6",
              "R7"
            ],
            "title": "Decision Code"
          },
          "count": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Count"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "decision_code",
          "count"
        ],
        "title": "ReportDecisionCount"
      },
      "ReportObjectRow": {
        "properties": {
          "object_id": {
            "type": "string",
            "maxLength": 256,
            "minLength": 1,
            "title": "Object Id"
          },
          "object_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Object Name"
          },
          "cards": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Cards"
          },
          "released": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Released"
          },
          "source_alarms": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Source Alarms"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "object_id",
          "cards",
          "released",
          "source_alarms"
        ],
        "title": "ReportObjectRow"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "integrationToken": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "ipk_...",
        "description": "Токен интеграции внешней системы (выдаёт администратор сервиса). Даёт только право ingest: отправку пачек и статус своих пачек."
      },
      "sessionCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "__Host-infrapulse-session",
        "description": "Сессия входа через каталог. Ставится при входе в веб-интерфейс этого же сайта и отправляется браузером сама; поле в «Authorize» заполнять не нужно."
      },
      "csrfToken": {
        "type": "apiKey",
        "in": "header",
        "name": "X-CSRF-Token",
        "description": "Для POST под сессией: значение cookie __Host-infrapulse-csrf. С токеном интеграции не нужен."
      }
    }
  }
}